feat: integrate widgetConfig to widget
